==> Part1.ipynb <==
import pandas as pd

basics_url="https://datasets.imdbws.com/title.basics.tsv.gz"
akas_url="https://datasets.imdbws.com/title.akas.tsv.gz"
ratings_url="https://datasets.imdbws.com/title.ratings.tsv.gz"

basics = pd.read_csv(basics_url, sep='\t', low_memory=False)
akas = pd.read_csv(akas_url, sep='\t', low_memory=False)
ratings = pd.read_csv(ratings_url, sep='\t', low_memory=False)



import numpy as np
basics = pd.read_csv(basics_url, sep='\t', low_memory=False)

basics=basics.replace({'\\N':np.nan})
ratings=ratings.replace({'\\N':np.nan})
akas=akas.replace({'\\N':np.nan})

akas = akas[akas['region'] == 'US']

basics=basics.dropna(subset=['genres', 'runtimeMinutes'])
basics = basics[basics['titleType'] == 'movie']
basics['startYear'] = basics['startYear'].fillna(0)
basics['startYear'] =basics['startYear'].astype(int)
basics = basics[(basics['startYear'] >= 2000) & (basics['startYear'] <= 2021)]
is_documentary = basics['genres'].str.contains('documentary',case=False)
basics= basics[~is_documentary]

keepers =basics['tconst'].isin(akas['titleId'])
basics = basics[keepers]
keepers2 =ratings['tconst'].isin(akas['titleId'])
ratings =ratings[keepers2]


ratings.info()

basics.info()

akas.info()

basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)
akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)

#basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
#basics.head()

#ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
#ratings.head()

#akas = pd.read_csv("Data/title_akas.csv.gz", low_memory = False)
#akas.head()
