==> Part2.ipynb <==
# Install tmdbsimple (only need to run once)
!pip install tmdbsimple

import json
with open('/Users/jakec/.secret/tmdb_api.json', 'r') as f:
    login = json.load(f)
## Display the keys of the loaded dict
login.keys()

import tmdbsimple as tmdb
tmdb.API_KEY =  login['client-id']

## make a movie object using the .Movies function from tmdb
movie = tmdb.Movies(603)
movie.info()
## movie objects have a .info dictionary 
#movie.info()


def get_movie_with_rating(movie_id):
    movie=tmdb.Movies(movie_id)
    info = movie.info()
    releases = movie.releases()
    for c in releases['countries']:
        if c['iso_3166_1']=='US':
            info['certification'] = c['certification']
    return info

test = get_movie_with_rating("tt0848228") #put your function name here
test



## testing our function by looping through a list of ids
import pandas as pd
test_ids = ["tt0848228", "tt0115937","tt0848228","tt0332280"]
results = []
for movie_id in test_ids:
    movie_info = get_movie_with_rating(movie_id)
    results.append(movie_info)
    
    
pd.DataFrame(results)



## testing our function by looping through a list of ids
import pandas as pd
test_ids = ["tt0848228", "tt0115937","tt0848228","tt0332280"]
results = []
for movie_id in test_ids:
    
    try:
        movie_info = get_movie_with_rating(movie_id)
        results.append(movie_info)
        
    except: 
        pass
    
pd.DataFrame(results)



## testing our function by looping through a list of ids
import pandas as pd
test_ids = ["tt0848228", "tt0115937","tt0848228","tt0332280"]
results = []
errors = []
for movie_id in test_ids:
    
    try:
        movie_info = get_movie_with_rating(movie_id)
        results.append(movie_info)
        
    except Exception as e: 
        errors.append([movie_id, e])
    
pd.DataFrame(results)



print(f"- Number of errors: {len(errors)}")
errors



def write_json(new_data, filename): 
    """Appends a list of records (new_data) to a json file (filename). 
    Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/"""  
    
    with open(filename,'r+') as file:
        # First we load existing data into a dict.
        file_data = json.load(file)
        ## Choose extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
             file_data.append(new_data)
        # Sets file's current position at offset.
        file.seek(0)
        # convert back to json.
        json.dump(file_data, file)


import pandas as pd
# Load in the dataframe from project part 1 as basics:
basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)

YEARS_TO_GET = [2000,2001]

errors = [ ]



from tqdm.notebook import tqdm_notebook
FOLDER = "Data/"
import os
# Start of OUTER loop
for YEAR in tqdm_notebook(YEARS_TO_GET, desc='YEARS', position=0):
    #Defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    # Check if file exists
    file_exists = os.path.isfile(JSON_FILE)
    # If it does not exist: create it
    if file_exists == False:
    # save an empty dict with just "imdb_id" to the new json file.
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)
    #Saving new year as the current df
    df = basics.loc[ basics['startYear']==YEAR].copy()
    # saving movie ids to list
    movie_ids = df['tconst'].copy()
    # Load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]
        #Get index and movie id from list
    # INNER Loop
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                  desc=f'Movies from {YEAR}',
                                  position=1,
                                  leave=True):
        try:
            # Retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)  
            # Append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)
            
        except Exception as e:
            errors.append([movie_id, e])
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)

print(f"- Total errors: {len(errors)}")


EDA

data2000 = pd.read_csv("Data/final_tmdb_data_2000.csv.gz", low_memory = False)
data2001 = pd.read_csv("Data/final_tmdb_data_2001.csv.gz", low_memory = False)

datacombined=pd.concat([data2000, data2001])

datacombined

How many movies had at least some valid financial information (values > 0 for budget OR revenue)?

((datacombined['budget'] > 0) ^ (datacombined['revenue'] > 0)).sum()

284 movies.

How many movies are there in each of the certification categories (G/PG/PG-13/R)?


datacombined['certification'].value_counts()

What is the average revenue per certification category?


datacombined.groupby('certification')['revenue'].mean()

What is the average budget per certification category?


datacombined.groupby('certification')['budget'].mean()

datacombined.to_csv("Data/tmdb_results_combined.csv.gz",compression='gzip',index=False)
