==> Part3.ipynb <==
import pandas as pd

basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics['genres_split'] = basics['genres'].str.split(',')
#deletelist1=
basics = basics.drop(['titleType', 'originalTitle','isAdult'], axis=1)
basics

ratings = pd.read_csv("Data/title_ratings.csv.gz", low_memory = False)
ratings

TMDBAPI = pd.read_csv("Data/tmdb_results_combined.csv.gz", low_memory = False)
TMDBAPI = TMDBAPI[['imdb_id', 'revenue', 'budget', 'certification']]
TMDBAPI = TMDBAPI.drop(TMDBAPI.index[0])
TMDBAPI= TMDBAPI.reset_index(drop=True)
TMDBAPI

exploded_genres = basics.explode('genres_split')
exploded_genres

unique_genres = sorted(exploded_genres['genres_split'].unique())

title_genres=exploded_genres[['tconst', 'genres_split']].copy()
title_genres = title_genres.rename(columns={'genres_split': 'genre_ID'})
title_genres.head()

## Making the genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

## make new integer genre_id and drop string genres
exploded_genres['genre_id'] = exploded_genres['genres_split'].map(genre_map)
exploded_genres = exploded_genres.drop(columns='genres_split')


genre_lookup=pd.DataFrame({'Genre_Name': genre_map.keys(), 'Genre_ID':genre_map.values()})
genre_lookup

basics = basics.drop(['genres', 'genres_split'], axis=1)
basics

import json
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists
with open('/Users/jakec/.secret/mySQL.json', 'r') as f:
    login = json.load(f)
## Display the keys of the loaded dict
login.keys()

username = 'root' # default username for MySQL db is root
password = 'Snakeking3!' # whatever password you chose during MySQL installation.

connection_str = f'mysql+pymysql://{username}:{password}@localhost/Movies'
engine = create_engine(connection_str)

# check if database exists, if not, create it
if database_exists(connection_str):
    print('It exists')
else:
    create_database(connection_str)
    print('Database created!')

from sqlalchemy.types import *
## Calculate max string lengths for object columns
key_len = basics['tconst'].fillna('').map(len).max()
title_len = basics['primaryTitle'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(key_len+1), 
    "primaryTitle": Text(title_len+1),
    'startYear':Float(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}



# Save to sql with dtype and index=False
basics.to_sql('title_basics',engine,dtype=df_schema,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

key_len2 = ratings['tconst'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema2 = {
    "tconst": String(key_len2+1), 
    "averageRating": Float(),
    'numVotes':Integer()}


ratings.to_sql('title_ratings',engine,dtype=df_schema2,if_exists='replace',index=False)

engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')


title_genres.to_sql('title_genres',engine,if_exists='replace',index=False)

genre_lookup.to_sql('genres', engine, if_exists='replace', index=False)


key_len3 = TMDBAPI['imdb_id'].fillna('').map(len).max()
title_len3 = TMDBAPI['certification'].fillna('').map(len).max()
## Create a schema dictonary using Sqlalchemy datatype objects
df_schema3 = {
    "imdb_id": String(key_len3+1), 
    "revenue": Float(),
    "Budget": Float(),
    'certification':String(title_len3+1)}


TMDBAPI.to_sql('tmdb_data',engine,dtype=df_schema3,if_exists='replace',index=False)

engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

q = """SELECT * FROM title_basics LIMIT 5;"""
pd.read_sql(q, engine)



q = """SELECT * FROM title_ratings LIMIT 5;"""
pd.read_sql(q, engine)



q = """SELECT * FROM title_genres LIMIT 5;"""
pd.read_sql(q, engine)



q = """SELECT * FROM genres LIMIT 5;"""
pd.read_sql(q, engine)



q = """SELECT * FROM tmdb_data LIMIT 5;"""
pd.read_sql(q, engine)

q = """SHOW TABLES;"""
pd.read_sql(q, engine)
