==> Part4.ipynb <==
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd
import numpy as np
from scipy import stats
TMDBAPI = pd.read_csv("Data/tmdb_results_combined.csv.gz", low_memory = False)
TMDBAPI = TMDBAPI[['imdb_id', 'revenue', 'budget', 'certification']]
TMDBAPI = TMDBAPI.drop(TMDBAPI.index[0])
TMDBAPI= TMDBAPI.reset_index(drop=True)
TMDBAPI

TMDBAPI2=TMDBAPI.copy()
TMDBAPI2=TMDBAPI2[TMDBAPI2['certification'].isin(['R', 'PG-13', 'PG', 'G'])]

TMDBAPI2['certification'].value_counts()

The stakeholder's first question is: does the MPAA rating of a movie (G/PG/PG-13/R) affect how much revenue the movie generates?

There are four categories to compare and we are dealing with numerical values, therefore, ANOVA method is used.

Null Hypothesis: There is no difference on revenue produced for ratings of a movie.

Alternate Hypothesis: There is an impact on revenue based on the ratings of a movie.

ratings ={}

for rating in TMDBAPI2['certification'].unique():
    temp = TMDBAPI2.loc[TMDBAPI2['certification']== rating, 'revenue']
    ratings[rating] = temp

ratings.keys()

for certification, rating in ratings.items():

    ## determine if there are any outliers
    outliers = np.abs(stats.zscore(rating)) > 3

    ## print a statement about how many outliers for which group name
    print(f"There were {outliers.sum()} outliers in the {certification} ratings.")

    ## Remove the outiers from data and overwrite the sector data in the dict
    rating = rating.loc[~outliers]
    ratings[certification] = rating

## Running normal test on each group and confirming there are >20 in each group

## Save a list with an inner list of column names
norm_results = [['ratings','n','pval','sig?']]


## loop through group dict
for certification, rating in ratings.items():
    ## calculate normaltest results
    stat, p = stats.normaltest(rating)

    ## Append the right info into norm_resutls (as a list)
    norm_results.append([certification,len(rating), p, p<.05])


## Make norm_results a dataframe (first row is columns, everything else data)
normal_results = pd.DataFrame(norm_results[1:], columns = norm_results[0])
normal_results

n is large enough to disregard this test.

stats.levene(*ratings.values())

We met the assumption of normality but did NOT meet the assumption of equal variance. We wanted to run an ANOVA test but will need to use a Kurskal-Wallis test instead.

result = stats.kruskal(*ratings.values())
print(result)
result.pvalue<.05

Our Kruskal Wallis test returned a p-value less than .05. We  reject the null hypothesis and conclude that there is a significant proof to say that the four groups have different revenue due to their movie ratings.

mean_values = TMDBAPI2.groupby('certification').mean()
mean_values

In average, PG-13 movies have the highest revenue while R rated movies have the lowest revenue.

The stakeholder's second question is: does the MPAA rating of a movie (G/PG/PG-13/R) affect the budget of the movie?

There are four categories to compare and we are dealing with numerical values, therefore, ANOVA method is used.

Null Hypothesis: There is no difference on budget produced for ratings of a movie.

Alternate Hypothesis: There is an impact on revenue based on the budget of a movie.

TMDBAPI3=TMDBAPI.copy()
TMDBAPI3=TMDBAPI3[TMDBAPI3['certification'].isin(['R', 'PG-13', 'PG', 'G'])]

ratings2 ={}

for rating in TMDBAPI3['certification'].unique():
    temp = TMDBAPI3.loc[TMDBAPI3['certification']== rating, 'budget']
    ratings2[rating] = temp

ratings2.keys()

for certification, rating in ratings2.items():

    ## determine if there are any outliers
    outliers = np.abs(stats.zscore(rating)) > 3

    ## print a statement about how many outliers for which group name
    print(f"There were {outliers.sum()} outliers in the {certification} ratings.")

    ## Remove the outiers from data and overwrite the sector data in the dict
    rating = rating.loc[~outliers]
    ratings2[certification] = rating

## Running normal test on each group and confirming there are >20 in each group

## Save a list with an inner list of column names
norm_results2 = [['ratings','n','pval','sig?']]


## loop through group dict
for certification, rating in ratings2.items():
    ## calculate normaltest results
    stat, p = stats.normaltest(rating)

    ## Append the right info into norm_resutls (as a list)
    norm_results2.append([certification,len(rating), p, p<.05])


## Make norm_results a dataframe (first row is columns, everything else data)
normal_results2 = pd.DataFrame(norm_results2[1:], columns = norm_results2[0])
normal_results2

n is large enough to disregard this test.

stats.levene(*ratings.values())

We met the assumption of normality but did NOT meet the assumption of equal variance. We wanted to run an ANOVA test but will need to use a Kurskal-Wallis test instead.

result = stats.kruskal(*ratings.values())
print(result)
result.pvalue<.05

Our Kruskal Wallis test returned a p-value less than .05. We reject the null hypothesis and conclude that there is a significant proof to say that the four groups have different budget based on their movie ratings.

mean_values2 = TMDBAPI3.groupby('certification').mean()
mean_values2

PG 13 movies tend to have the highest budget while the rated R movies tend to have the lowest budget. 

The stakeholder's third question is: Did the running time of movies change over the years?

basics = pd.read_csv("Data/title_basics.csv.gz", low_memory = False)
basics['startYear'].value_counts()

There are 20 different years to compare and we are dealing with numerical values, therefore, ANOVA method is used.

Null Hypothesis: There is no difference in running time over the years.

Alternate Hypothesis: There is an change in running time over the years.

years ={}

for year in basics['startYear'].unique():
    temp = basics.loc[basics['startYear']== year, 'runtimeMinutes']
    years[year] = temp

years.keys()

for year, data in years.items():

    ## determine if there are any outliers
    outliers = np.abs(stats.zscore(data)) > 3

    ## print a statement about how many outliers for which group name
    print(f"There were {outliers.sum()} outliers in the year {year}.")

    ## Remove the outiers from data and overwrite the sector data in the dict
    data = data.loc[~outliers]
    years[year] = data

norm_results = [['group','n','pval','sig?']]


## loop through group dict
for year, data in years.items():
    ## calculate normaltest results
    stat, p = stats.normaltest(data)

    ## Append the right info into norm_resutls (as a list)
    norm_results.append([year,len(data), p, p<.05])


## Make norm_results a dataframe (first row is columns, everything else data)
normal_results = pd.DataFrame(norm_results[1:], columns = norm_results[0])
normal_results

This shows that the groups do not have normal distribution, but our n is large enough to disregard it.



stats.levene(*years.values())


We met the assumption of normality but did NOT meet the assumption of equal variance. We wanted to run an ANOVA test but will need to use a Kurskal-Wallis test instead.



result = stats.kruskal(*years.values())
print(result)
result.pvalue<.05

Our Kruskal Wallis test returned a p-value less than .05. We reject the null hypothesis and conclude that there is a significant change over the years in runtime.
